+++ src/components/Animal Pages/Animal.js
