+++ src/components/TrainingPages/TrainingForm.js
