+++ src/pages/alltraining.js
@@ -2,9 +2,9 @@ import Image from "next/image";
 import { Inter } from "next/font/google";
 import quarterCircle from '/public/images/quarterCircle.png';
 import Navbar from "@/components/Navbar/Navbar.js";
-import SideBar from "@/components/Sidebar";
-import TopBanner from "@/components/TopBanner";
-import TrainingLog from "@/components/TrainingLog";
+import SideBar from "@/components/Sidebar/Sidebar";
+import TopBanner from "@/components/TopBanner/TopBanner";
+import TrainingLog from "@/components/TrainingPages/TrainingLog";
 
 
 export default function allTraining() {
