+++ src/pages/allusers.js
@@ -1,10 +1,10 @@
 import Image from "next/image";
 import { Inter } from "next/font/google";
 import Navbar from "@/components/Navbar/Navbar.js";
-import SideBar from "@/components/Sidebar";
-import TrainingLog from "@/components/TrainingLog";
-import TopBanner from "@/components/TopBanner";
-import AllUsersContainer from "@/components/allUsersContainer"
+import SideBar from "@/components/Sidebar/Sidebar";
+import TrainingLog from "@/components/TrainingPages/TrainingLog";
+import TopBanner from "@/components/TopBanner/TopBanner";
+import AllUsersContainer from "@/components/UsersPage/allUsersContainer"
 
 
 export default function allUsers() {
